<Refactor> Utils 패키지 생성 후 InputValidator, NumberGenerator 해당 패키지로 이동
