+++ src/main/java/racingcar/Utils/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.Utils;
 
 import camp.nextstep.edu.missionutils.Randoms;
